Apply clang-format fixes to my task

+++ tasks/kazenova_a_vec_change_sign/tests/performance/main.cpp
@@ -1,5 +1,7 @@
 #include <gtest/gtest.h>
 
+#include <cstddef>
+
 #include "kazenova_a_vec_change_sign/common/include/common.hpp"
 #include "kazenova_a_vec_change_sign/mpi/include/ops_mpi.hpp"
 #include "kazenova_a_vec_change_sign/seq/include/ops_seq.hpp"
@@ -7,9 +9,8 @@
 
 namespace kazenova_a_vec_change_sign {
 
-class KazenovaAVecChangeSignPerfTest
-    : public ppc::util::BaseRunPerfTests<InType, OutType> {
-protected:
+class KazenovaAVecChangeSignPerfTest : public ppc::util::BaseRunPerfTests<InType, OutType> {
+ protected:
   void SetUp() override {
     int vector_size = 100000000;
     input_data_.resize(vector_size);
@@ -19,12 +20,14 @@ protected:
   }
   bool CheckTestOutputData(OutType &output_data) final {
     size_t max_possible = input_data_.size() - 1;
-    return output_data >= 0 && static_cast<size_t>(output_data) <= max_possible;
+    return output_data >= 0 && output_data <= static_cast<int>(max_possible);
   }
 
-  InType GetTestInputData() final { return input_data_; }
+  InType GetTestInputData() final {
+    return input_data_;
+  }
 
-private:
+ private:
   InType input_data_;
 };
 
@@ -32,16 +35,13 @@ TEST_P(KazenovaAVecChangeSignPerfTest, RunPerfModes) {
   ExecuteTest(GetParam());
 }
 
-const auto kAllPerfTasks =
-    ppc::util::MakeAllPerfTasks<InType, KazenovaAVecChangeSignMPI,
-                                KazenovaAVecChangeSignSEQ>(
-        PPC_SETTINGS_kazenova_a_vec_change_sign);
+const auto kAllPerfTasks = ppc::util::MakeAllPerfTasks<InType, KazenovaAVecChangeSignMPI, KazenovaAVecChangeSignSEQ>(
+    PPC_SETTINGS_kazenova_a_vec_change_sign);
 
 const auto kGtestValues = ppc::util::TupleToGTestValues(kAllPerfTasks);
 
 const auto kPerfTestName = KazenovaAVecChangeSignPerfTest::CustomPerfTestName;
 
-INSTANTIATE_TEST_SUITE_P(RunModeTests, KazenovaAVecChangeSignPerfTest,
-                         kGtestValues, kPerfTestName);
+INSTANTIATE_TEST_SUITE_P(RunModeTests, KazenovaAVecChangeSignPerfTest, kGtestValues, kPerfTestName);
 
-} // namespace kazenova_a_vec_change_sign
+}  // namespace kazenova_a_vec_change_sign
